catch simulation, with/without protocol

+++ attack_data/jitter/experiment.sh
@@ -15,7 +15,7 @@ HPING_PORT=5555  # Port used by hping3
 # Remove any existing .pcap files from the drone1 pod before starting the experiment
 kubectl exec  drone1 --  rm -f *.pcap
 
-declare -a HPING_OPTIONS=("fast" "faster" "flood")
+declare -a HPING_OPTIONS=("flood")
 
 for opt in "${HPING_OPTIONS[@]}"; do
     for i in {1..3}; do
@@ -32,8 +32,10 @@ for opt in "${HPING_OPTIONS[@]}"; do
         sleep 10
         
         # Start hping3 on drone3 to send packets to drone1
-        kubectl exec  drone3 -- hping3 "$(kubectl get pod drone1 -o jsonpath='{.status.podIP}')" --udp -p $HPING_PORT "--$opt" &>/dev/null &
-        pid_hping3=$!
+	if [ $i -eq 2 ]; then
+	        kubectl exec  drone3 -- hping3 "$(kubectl get pod drone1 -o jsonpath='{.status.podIP}')" --udp -p $HPING_PORT "--$opt" &>/dev/null &
+	        pid_hping3=$!
+	fi
 
         # Wait for the specified duration before stopping the tcpdump process
         sleep $DURATION
@@ -46,7 +48,9 @@ for opt in "${HPING_OPTIONS[@]}"; do
         kubectl exec  drone2 -- pkill ffmpeg
 
         # Stop hping3 on drone3
-        kubectl exec  drone3 -- pkill hping3
+	if [ $i -eq 2 ]; then
+	        kubectl exec  drone3 -- pkill hping3
+	fi
 
         # Download the output file to the local machine
         kubectl cp drone1:$OUTPUT_FILE "${PREFIX}_tcpdump_output_${opt}_${i}.pcap"
@@ -55,6 +59,6 @@ for opt in "${HPING_OPTIONS[@]}"; do
         kubectl exec  drone1 -- rm -f $OUTPUT_FILE
 
         # Sleep for 50 seconds before starting the next iteration
-        sleep 50
+        sleep 10
     done
 done
